requirements.txt
playwright
pytest
pytest-xdist
allure-pytest
pyyaml
python-dotenv